requirements.txt
fastapi
tinydb
uvicorn
opencv-python
mediapipe
numpy
Pillow